requirements.txt
numpy
pandas

torch
torchvision

matplotlib
seaborn
scikit-learn

fastapi
uvicorn[standard]
python-multipart

Pillow
tqdm
python-dotenv
